requirements.txt
python-telegram-bot
urlextract